### 2025-03-17 #7912 @editorialbot
[PRE REVIEW]: substrateminer: A Python package to investigate protein substrate repertoires
**Submitting author:** @manifestoso (<a href="http://orcid.org/0000-0003-4569-6921">Robert QIAO</a>)
**Repository:** https://github.com/DPP4ResearchGroup/substrateminer.git
**Branch with paper.md** (empty if default branch):
**Version:** pd.v1.0
