### 2025-03-25 29bce62
HOTFIX - BibTex - fix #64
