+++ paper/paper.bib
@@ -22,17 +22,15 @@
 }
 
 @article{osti_6943966,
-title = {Ethanol from biomass by enzymatic hydrolysis},
-author = {Wright, J D},
-abstractNote = {Enzymes are biological catalysts that generally are designed to do one job well, but to do one job only. Therefore, the enzymes that catalyze the hydrolysis of cellulose to sugar do not break down the sugars. Thus, enzymatic processes are capable of yields approaching 100%. Enzymatic hydrolysis processes have been under development for only 10 years. Although improvements have been made in enzymatic technology, more are both possible and necessary. The important research issues include understanding the processes necessary to render the crystalline cellulose easily digestible, understanding and improving the basic mechanisms in the hydrolysis step, and developing better and less expensive enzymes. The hemicellulose fraction (25%) is primarily composed of xylan, which is simple to convert to the simple sugar xylose, but the xylose is difficult to ferment to ethanol. There were no practical systems for xylose fermentation 10 years ago. Today, methods have been identified using new yeasts, fungi, bacteria, and processes combining enzymes and yeasts. Although none of the fermentations is yet ready for commercial use, considerable progress has been made. The following sections describe current research efforts in each of the major areas (cellulose hydrolysis, xylose fermentation, and lignin conversion), with an emphasis on enzymatic hydrolysis using fungal enzymes.},
-doi = {},
-url = {https://www.osti.gov/biblio/6943966}, journal = {Chem. Eng. Prog.; (United States)},
-number = ,
-volume = 84:8,
-place = {United States},
-year = {1988},
-month = {8}
-}
+  author       = {Wright, J D},
+  title        = {Ethanol from biomass by enzymatic hydrolysis},
+  annote       = {Enzymes are biological catalysts that generally are designed to do one job well, but to do one job only. Therefore, the enzymes that catalyze the hydrolysis of cellulose to sugar do not break down the sugars. Thus, enzymatic processes are capable of yields approaching 100%. Enzymatic hydrolysis processes have been under development for only 10 years. Although improvements have been made in enzymatic technology, more are both possible and necessary. The important research issues include understanding the processes necessary to render the crystalline cellulose easily digestible, understanding and improving the basic mechanisms in the hydrolysis step, and developing better and less expensive enzymes. The hemicellulose fraction (25%) is primarily composed of xylan, which is simple to convert to the simple sugar xylose, but the xylose is difficult to ferment to ethanol. There were no practical systems for xylose fermentation 10 years ago. Today, methods have been identified using new yeasts, fungi, bacteria, and processes combining enzymes and yeasts. Although none of the fermentations is yet ready for commercial use, considerable progress has been made. The following sections describe current research efforts in each of the major areas (cellulose hydrolysis, xylose fermentation, and lignin conversion), with an emphasis on enzymatic hydrolysis using fungal enzymes.},
+  url          = {https://www.osti.gov/biblio/6943966},
+  journal      = {Chem. Eng. Prog.; (United States)},
+  volume       = {84:8},
+  place        = {United States},
+  year         = {1988},
+  month        = {08}}
 
 @article{GUO2023128252,
 title = {Enzymes and enzymatic mechanisms in enzymatic degradation of lignocellulosic biomass: A mini-review},
