### 2025-03-25 4ca212e
HOTFIX - BibTex - fix #64
