+++ paper/paper.bib
@@ -39,7 +39,7 @@ volume = {367},
 pages = {128252},
 year = {2023},
 issn = {0960-8524},
-doi = {https://doi.org/10.1016/j.biortech.2022.128252},
+doi = {10.1016/j.biortech.2022.128252},
 url = {https://www.sciencedirect.com/science/article/pii/S0960852422015851},
 author = {Hongliang Guo and Ying Zhao and Jo-Shu Chang and Duu-Jong Lee},
 keywords = {Lignocellulose, Hydrolysis, Mechanism, Kinetics},
@@ -205,7 +205,7 @@ volume = {28},
 number = {11},
 pages = {1947-1951},
 keywords = {KEGG, KEGG MEDICUS, KEGG module, pathway mapping, reaction module},
-doi = {https://doi.org/10.1002/pro.3715},
+doi = {10.1002/pro.3715},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/pro.3715},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/pro.3715},
 abstract = {Abstract In this era of high-throughput biology, bioinformatics has become a major discipline for making sense out of large-scale datasets. Bioinformatics is usually considered as a practical field developing databases and software tools for supporting other fields, rather than a fundamental scientific discipline for uncovering principles of biology. The KEGG resource that we have been developing is a reference knowledge base for biological interpretation of genome sequences and other high-throughput data. It is now one of the most utilized biological databases because of its practical values. For me personally, KEGG is a step toward understanding the origin and evolution of cellular organisms.},
