### 2025-05-21 #7912 @csoneson
Hi @manifestoso - I will handle your submission and start by finding at least two suitable reviewers. If you have suggestions (e.g. from the list mentioned in the first post above), feel free to mention them here (but please don't tag them).

Regarding the question about the `Statement of need` section - yes, this should be part of the paper. Please see [here](https://joss.readthedocs.io/en/latest/paper.html) for the JOSS paper guidelines.

@tushardave26 - are you still available to review this submission?
