### 2025-05-24 #8266 @csoneson
@manifestoso - please also take a look at the paper guidelines pointed to in the [pre-review issue](https://github.com/openjournals/joss-reviews/issues/7912#issuecomment-2896919412), e.g. regarding the `Statement of need` section.
