### 2025-06-17 #8266 @csoneson
> 👋🏻 Hi all - just wanted to check in to see how things are going and if there are any questions about the process.
>
> [@manifestoso](https://github.com/manifestoso) - please have a look at adding the `Statement of need` section to the paper (you can then generate a new proof with `@editorialbot generate pdf`) so that the reviewers can evaluate this as well.

👋🏻 Hi all - could you give me an update on how things are going here? @tushardave26 - I note that you didn't yet create your checklist, let me know if anything is unclear. @manifestoso - as mentioned, please have a look at adding the Statement of need section to the paper.
