### 2025-06-27 DPP4ResearchGroup/substrateminer#73 @tushardave26
Review Comments from JOSS Reviewer (@tushardave26)
Hi @manifestoso, as part of my review of the JOSS submission (https://github.com/openjournals/joss-reviews/issues/8266), I’ve outlined several suggestions aimed at improving and enhancing the package. These are organized into two categories: (1) **Mandatory** recommendations and (2) **Optional** enhancements.

Please note that this is an ongoing review. I’ve initialized this issue with some preliminary observations and will continue to add more feedback as I progress through the review process.

**Mandatory Recommendations**

- [x] Add **Statement of Need** to the manuscript.
- [x] Change the `default` branch to `main` branch.
- [x] Fix the failure of `unittest` on the `main` branch. (Refer to the `UnitTest Status` section for this)
- [x] Rewrite the installation section to clearly state that users must clone the repository before installing via `conda`, as the installation relies on local `YAML` files.
- [x] Correct the `conda` activation `CLI`. After initializing the `conda` `env`, `conda` suggested to activate the `substrateminer` with the following command line and not the one indicated in the `README` or the `documentation`.
conda activate substrate-miner
```
- [x] Fix the deprecation warning for the `Bio.Application` module. See the deprecation warning below while I tried running the `substrateminer --help` command.
```
/usr/local/anaconda3/envs/substrate-miner/lib/python3.10/site-packages/Bio/Application/__init__.py:39: BiopythonDeprecationWarning: The Bio.Application modules and modules relying on it have been deprecated.

Due to the on going maintenance burden of keeping command line application
wrappers up to date, we have decided to deprecate and eventually remove these
modules.

We instead now recommend building your command line and invoking it directly
with the subprocess module.
  warnings.warn(
```
- [x] Fix the issue https://github.com/DPP4ResearchGroup/substrateminer/issues/70
- [x] Replace `,` with `.` in the first statement of the `Quick Guide` section in the `repo` and `documentation` as well.
- [x] Consider revising the statement [here](https://github.com/DPP4ResearchGroup/substrateminer/tree/develop?tab=readme-ov-file#installation), as it currently contains a spelling error in the word “recommend” and could benefit from more precise phrasing.
- [x] Add `CONTRIBUTE.md` file to the repo. I see that you have a section for this in the `documentation`. Please add that to the repository as a separate file.
- [x] Consider combining the two sections of the `Installation`. The `README` of the repository and the `documentation` both contain two sections of the `Installation`.
- [x] Consider rephrasing the version requirements for clarity and consistency. For example, `Python >= 3.10.11`, `BioPython >= 1.84`, `matplotlib <= 3.6.0`.
- [x] **Manuscript Changes**
    - [x] Add benchmarking or comparison with existing tools or workflows, if any.
    - [x] Add a limitation section - Acknowledge that consensus-based mining may miss non-canonical substrates or novel cleavage mechanisms.
    - [x] Clarify KEGG usage scope - Is it limited to human proteins or applicable across species?
    - [x] Fix typo with `KEGG accession conversion enables a gateway` at Line 67

**Optional Enhancements**

- [x] Consider adding an alternative installation option that does not rely on `conda`, such as direct install using `pip`.
- [x] Consider switching the order of the sections in both the `README` and the `documentation`. The flow would be more intuitive if the requirements and installation instructions appeared first, followed by the quick start guide and other sections.
