### 2025-07-19 #8266 @csoneson
Ping @manifestoso

> 👋🏻 [@manifestoso](https://github.com/manifestoso) Just checking in here to see where we're at with the revisions. Let us know if anything is unclear.
