### 2025-07-23 DPP4ResearchGroup/substrateminer#73 @manifestoso
Regarding Optional change 1:

In term of packaing, `conda` is more practical at the moment largely to due two alignment binary utilities required, install those from source prob is less user-friendly.  Alternative, we can package this utility as a docker distribution in the next release. I will mark this off for the time being, feel free to bounce any ideas. Cheers
