### 2025-07-24 33af12f
JOSS PeerR fix - DPP4ResearchGroup#73, DPP4ResearchGroup#75, doco fix - [x] inc HOTFIX - typo in sec Installation
