+++ docs/README.md
@@ -1,17 +1,78 @@
 # substrateminer
+
+[![status](https://joss.theoj.org/papers/09b68390d13060894366a1626b45aa05/status.svg)](https://joss.theoj.org/papers/09b68390d13060894366a1626b45aa05)
+
 `substrateminer` is a python package that offer a suite of discovery tools to investigate enzyme substrate repertorie based on sequence cleavage consensus.
 
 ## CI/CD Status
+
 ### UnitTest Status
 
-| Branch | `main` | `develop` | `features` |
-|:--------|:--------|:--------|:--------|
-|Linux|![substrateminer-main](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/python-ci.yml/badge.svg?branch=main)|![substrateminer-dev](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/python-ci.yml/badge.svg?branch=develop)|![substrateminer-features](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/python-ci.yml/badge.svg?branch=features)|
-|macOS| [![substrateminer-main](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml/badge.svg?branch=main)](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml)| [![substrateminer-dev](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml/badge.svg?branch=develop)](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml) | [![substrateminer-features](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml/badge.svg?branch=features)](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/substrateminer-mac.yml) |
+| Branch | `main`                                                                                                                                                                                                                                     | `develop`                                                                                                                                                                                                                                    | `features`                                                                                                                                                                                                                                         |
+|:-------|:-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|:---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
[…]
