+++ README-index.md
@@ -1,0 +1,22 @@
+---
+absorb: true
+---
+
+- [substrateminer](readme_docs/header.md)
+    - [Overview](readme_docs/summary.md)
+    - CI/CD Status
+        - [UnitTest Status](readme_docs/ci_status.md)
+        - [Documentation Status](readme_docs/gh_status.md)
+    - TL;DR
+        - [Installation Guide](readme_docs/installation.md)
+            - [Requirements](readme_docs/requirements.md)
+        - [Quick Start Guide](readme_docs/quick_start.md)
+            - [Usage Examples](readme_docs/examples.md)
+    - [Construct a customised workflow](readme_docs/custom_workflow.md)
+        - [Methods and Functions Overview](readme_docs/methods.md)
+    - GitHub Actions CI manual
+        - [UnitTests Sequence](readme_docs/ci_sequence.md)
+    - [License](readme_docs/license.md)
+    - [Issue/Bug Reporting](readme_docs/issue_reporting.md)
+    - [Contributing](readme_docs/contributing.md)
+       
