+++ readme_docs/quick_start.md
@@ -1,0 +1,17 @@
+`substrateminer` provides three main categories of functionalities namely `motif`, `miner`, and `pathfinder`. `substrateminer` also integrates multi-sequence alignment tools to facilitate the analysis.
+```
+Usage: substrateminer [OPTIONS] COMMAND [ARGS]...
+
+  A suite to tools to discover enzyme substrates based on sequence consensus.
+
+  Main entry point for substrateminer CLI.
+
+Options:
+  --help  Show this message and exit.
+
+Commands:
+  miner       Filter amino acid sequences from a reference file.
+  motif       Interface for consensus sequence determination.
+  msa         Interface for multi-sequence alignments
+  pathfinder  Find the pathological/molecular path for a substrate.
+```
