+++ readme_docs/ci_sequence.md
@@ -1,0 +1,8 @@
+CI/CD is carried out with GitHub Actions workflow and consists following steps:
+
+- Checks out the repository.
+- Sets up Python.
+- Caches the conda environment.
+- Installs Miniconda and creates the conda environment.
+- Runs CLI tests with pip.
+- Runs unit tests with unittest.
