+++ readme_docs/summary.md
@@ -1,0 +1,4 @@
+
+[![status](https://joss.theoj.org/papers/09b68390d13060894366a1626b45aa05/status.svg)](https://joss.theoj.org/papers/09b68390d13060894366a1626b45aa05)
+
+`substrateminer` is a python package that offer a suite of discovery tools to investigate enzyme substrate repertorie based on sequence cleavage consensus.
