+++ readme_docs/gh_status.md
@@ -1,0 +1,3 @@
+| Page | Status |
+|:--------|:--------|
+|substrateminer|[![pages-build-deployment](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/pages/pages-build-deployment/badge.svg)](https://github.com/DPP4ResearchGroup/substrateminer/actions/workflows/pages/pages-build-deployment)

[…]
