### 2025-07-24 bac3f17
JOSS PeerR - fix DPP4ResearchGroup#70, DPP4ResearchGroup#73, DPP4ResearchGroup#75: func fix
