+++ substrate_miner/msa/msa.py
@@ -1,6 +1,16 @@
-from Bio import SeqIO
-from Bio.Align.Applications import ClustalOmegaCommandline, MafftCommandline, MuscleCommandline
+#######
+# ChangeLog:
+# - Updated to deprecate Bio.Align.Applications for ClustalOmega, MAFFT, and MUSCLE.
+# - Added subprocess calls for these tools.
+#######
 
+from Bio import SeqIO
+#######
+# Align.Applications is now deprecated in Biopython 1.80+.
+# from Bio.Align.Applications import ClustalOmegaCommandline, MafftCommandline, MuscleCommandline
+#######
+
+import subprocess
 import argparse
 import click
 import sys
@@ -42,17 +52,44 @@ def perform_msa(input_file, output_file, method):
             os.remove(output_file)
 
         # Perform multiple sequence alignment using the specified method
-        if method == "clustalomega":
-            clustalomega_cline = ClustalOmegaCommandline(infile=input_file,\
-                outfile=output_file, verbose=True, auto=True)
-            stdout, stderr = clustalomega_cline()
+        if method == "clustal":
+            #######
+            # Change to subprocess call for ClustalOmega, as Bio.Align.Applications is deprecated
+            # clustalomega_cline = ClustalOmegaCommandline(infile=input_file,\
+            #    outfile=output_file, verbose=True, auto=True)
+            if sys.platform != "darwin":
+                clustalomega_exe = shutil.which("clustalo")
+            else:
+                clustalomega_exe = shutil.which("clustalw")
+                
+            if clustalomega_exe is None:
+                raise FileNotFoundError("ClustalOmega executable not found.\
+                    Please ensure it is installed and in your PATH.")
+            clustalomega_cline = subprocess.run([clustalomega_exe, "-i", input_file, "-o", output_file, "--force",\
+                "--verbose"], capture_output=True, text=True)
+            stdout, stderr = clustalomega_cline.stdout, clustalomega_cline.stderr
         elif method == "mafft":
             mafft_exe = shutil.which("mafft")
             if mafft_exe is None:
                 raise FileNotFoundError("MAFFT executable not found.\
                     Please ensure it is installed and in your PATH.")
-            mafft_cline = MafftCommandline(mafft_exe, input=input_file)
-            stdout, stderr = mafft_cline()
+            
+            #######
+            # Change to subprocess call for MAFFT, as Bio.Align.Applications is deprecated
+            # # mafft_cline = MafftCommandline(mafft_exe, input=input_file)
+            # stdout, stderr = mafft_cline()
+            # with open(output_file, "w") as mafft_w_handle:
+            #     mafft_w_handle.write(stdout)
+            #######
+            
+            # Use subprocess to call MAFFT
+            # Note: The input file should be in FASTA format
+            mafft_cline = subprocess.run([mafft_exe, input_file], capture_output=True, text=True)
+            stdout, stderr = mafft_cline.stdout, mafft_cline.stderr
+            # Write the output to the specified output file
+            if not stdout:
+                raise ValueError("MAFFT did not produce any output. Check the input file format.")
+            # Write the alignment results to the output file
             with open(output_file, "w") as mafft_w_handle:
                 mafft_w_handle.write(stdout)
         elif method == "muscle":
@@ -60,9 +97,24 @@ def perform_msa(input_file, output_file, method):
             if muscle_exe is None:
                 raise FileNotFoundError("MUSCLE executable not found.\
                     Please ensure it is installed and in your PATH.")
-            muscle_cline = MuscleCommandline(muscle_exe, input=input_file,\
-                out=output_file)
-            stdout, stderr = muscle_cline()
+            
+            #######
[…]
