### 2025-07-24 79e3c75
JOSS PeerR HOTFIX - DPP4ResearchGroup#73, DPP4ResearchGroup#75: doco md rendering update
