+++ Readme.md
@@ -43,7 +43,7 @@ OR
 ```
 # For MacOS users
 $ cd substrateminer
-$ conda env create -f environment-MacOS.yml # MacOS support
+$ conda env create -f environment-macOS.yml # MacOS support
 $ conda activate substrateminer
 ```
 
