### 2025-07-29 68239b4
JOSS PeerR itern##2 - fix DPP4ResearchGroup#80, fix DPP4ResearchGroup#69, fix DPP4ResearchGroup#75, fix DPP4ResearchGroup#73 - Paper hotfix - [x] HOTFIX - PDF compile
