+++ paper/paper.bib
@@ -212,4 +212,152 @@ abstract = {Abstract In this era of high-throughput biology, bioinformatics has 
 year = {2019}
 }
 
+@article{CaescuBJ20090549,
+    author = {Caescu, Cristina I. and Jeschke, Grace R. and Turk, Benjamin E.},
+    title = {Active-site determinants of substrate recognition by the metalloproteinases TACE and ADAM10},
+    journal = {Biochemical Journal},
+    volume = {424},
+    number = {1},
+    pages = {79-88},
+    year = {2009},
+    month = {10},
+    abstract = {The metalloproteinases TACE [tumour necrosis factor α-converting enzyme; also known as ADAM17 (a disintegrin and metalloproteinase 17)] and ADAM10 are the primary enzymes responsible for catalysing release of membrane-anchored proteins from the cell surface in metazoan organisms. Although the repertoire of protein substrates for these two proteases is partially overlapping, each one appears to target a subset of unique proteins in vivo. The mechanisms by which the two proteases achieve specificity for particular substrates are not completely understood. We have used peptide libraries to define the cleavage site selectivity of TACE and ADAM10. The two proteases have distinct primary sequence requirements at multiple positions surrounding the cleavage site in their substrates, which allowed us to generate peptide substrates that are highly specific for each of these proteases. The major difference between the two protease specificities maps to the P1′ position (immediately downstream of the cleavage site) of the substrate. At this position, TACE is selective for smaller aliphatic residues, whereas ADAM10 can accommodate aromatic amino acids. Using mutagenesis we identified three residues in the S1′ pockets of these enzymes that dramatically influence specificity for both peptide and protein substrates. Our results suggest that substrate selectivity of TACE and ADAM10 can be at least partly rationalized by specific features of their active sites.},
+    issn = {0264-6021},
+    doi = {10.1042/BJ20090549},
+    url = {https://doi.org/10.1042/BJ20090549},
+    eprint = {https://portlandpress.com/biochemj/article-pdf/424/1/79/659760/bj4240079.pdf},
+}
+
+@misc{kegg_statistics_2025,
+  title = {{KEGG} Database Current Statistics},
+  author = {Kanehisa, Minoru and Goto, Susumu},
+  year = {2025},
+  url = {https://www.genome.jp/kegg/docs/statistics.html},
+  urldate = {2025-07-28},
+  note = {Current statistics for KEGG (Kyoto Encyclopedia of Genes and Genomes) databases, accessed July, 2025},
+  organization = {Kanehisa Laboratories}
+}
+
+@article{BIGWARFE2004185,
+title = {Effect of ionization mode in the analysis of proteolytic protein digests},
+journal = {International Journal of Mass Spectrometry},
+volume = {234},
+number = {1},
+pages = {185-202},
+year = {2004},
+note = {Special Issue in Honour of Alan G. Marshall},
+issn = {1387-3806},
+doi = {https://doi.org/10.1016/j.ijms.2004.02.015},
+url = {https://www.sciencedirect.com/science/article/pii/S1387380604001150},
+author = {Paul M Bigwarfe and Troy D Wood},
+keywords = {Bottom-up, Ionization, Mass fingerprint},
[…]
