+++ paper/paper.md
@@ -21,34 +21,57 @@ bibliography: paper.bib
 ---
 
 # Abstract
-`Substrateminer` is a Python package that provides a suite of investigative tools to develop protein substrate repertoires for a given protease of interest.  `Substrateminer` provides tools for analysing protease cleavage consensus and visualising the conservation and strength of consensus. `Substrateminer` is a complete suite that enables potential substrate identification for a protease of interest in a target organism and extended functions allow for further discovery of associated biological and pathological functions of detected substrates.
+`Substrateminer` is a Python library that provides a suite of modular investigative methods that are based on protein sequence and cellular properties to support rapid development of *in silico* pipelines to discover and interrogate potential substrates for a target protease of interest. Three categories of methods included in `substrateminer` are consensus analysis, substrate mining, and pathological investigation, which together provide a well-suited toolkit to explore substrate repertoires also known as substrate degradome for a target protease.
 
-# Introduction
-Enzyme hydrolysis is a fundamental biochemical process in biology, and underpins many biological processes in nature from food digestion in humans [@campbell2005biology] to renewable energy production with cellulosic ethanol [@GUO2023128252;@osti_6943966]. Proteases are a class of enzymes that carry out such hydrolytic functions in biology by cleaving peptide bonds. In human physiology, a wide variety of critically important biophysiological processes depend on well-regulated protease functions, including protein degradation, cell signalling, and immune response. [@R800035200;@Bond2019;@ajplung001562016] Consequently, the inhibition of target proteases has long been one intervention strategy applied clinically to treat diverse pathophysiological conditions including viral infections like HIV, hepatitis-C; metabolic dysfunctions like type 2 diabetes [@Scott2017]; and increasingly, cancers [@Manasanch2017]. In the wake of the COVID-19 pandemic and increased respiratory diseases, proteases have emerged as a major therapeutic target for curbing fatality due to viral infections.[@ijms25158105;@v16030366;@ijms22115762]
+# Statement of need
+The ability to identify and discover novel protease *in situ* substrates is often a powerful first step towards understanding the biological functions of proteases and their implications in human health and disease. However, in traditional enzymology, substrates of a target enzyme are studied in a one-to-one relationship, arguably due to prohibitive cost involved in designing and deploying experimental assays with large quantity. Modern proteomics methods particularly the advancement in mass spectrometry, drastically reduce the cost associated with detecting substrate hydrolysis events at molecular level and now the study of substrates of a target enzyme at proteome level is no longer a distant fantasy. With these new possibilities, the lack of a modular system to enable rapid development of *in silico* pipeline to support preliminary discovery has been a bottleneck for new studies. Therefore, the introduction of `substrateminer` here aims to provide a versatile toolkit that enables fast modular development of *in silico* pipelines to facilitate the preliminary discovery that complement and accelerate the emerging experimental efforts.
 
-Nevertheless, the protease network in many organisms especially mammals including humans is complex and heavily-entangled, and the regulatory scope for many target proteases is still not well-understood. The ability to gather and identify novel protease substrates is an important step towards a better understanding of the biological functions of proteases and accelerate future developments of new protease inhibitors for downstream therapeutic applications. Here, the introduction of `substrateminer` aims to provide a comprehensive suite of tools to streamline the investigation and develop protein substrate repertoires for a given protease of interest based on a consensus algorithm.
+# Background
[…]
