### 2025-07-29 cc61836
JOSS PeerR ## 2 - HOTFIX - typo & stack cleanup
