+++ paper/paper.md
@@ -29,7 +29,7 @@ The ability to identify and discover novel protease *in situ* substrates is ofte
 # Background
 Proteolytic cleavage is a fundamental biochemical process in biology, which underpins many important biological processes that surround us from common food processing and digestion [@campbell2005biology] to cutting-edge industrial renewable energy production of cellulosic ethanol [@GUO2023128252;@osti_6943966]. Almost all proteolytic processes in nature are catalysed by enzymes and proteases are such an enzyme class that facilitates the biochemical reaction of cleaving peptide bonds. The inhibition of proteases has been a major intervention strategy in modern clinics to modulate molecular proteolytic processes to treat many physiological conditions including viral infections like HIV, hepatitis-C; metabolic dysfunctions like type-2 diabetes [@Scott2017]; and cancers [@Manasanch2017]. In the wake of the COVID-19 pandemic, proteases have emerged as a major therapeutic target for curbing fatality due to viral infections.[@ijms25158105;@v16030366;@ijms22115762]
 
-Nevertheless, one constraining factor that confronts the rapid development of protease inhibitory therapy is the undiscriminating nature of inhibition. This is a major challenge because most proteases are involved in a complex network of biological processes and a complete inhibition of one protease can lead to unintended consequences on many other biological processes. Therefore, identifying and understanding the substrate degradome of a target protease is an important preliminary step towards inhibitory therapy. Traditionally, the identification of protease substrates is often challenging due to the prohibitive cost of large scale deployment of target-designed immuno-based biochemistry assays and scarcity of adequate animal models. Modern advancements in proteomics, particularly the maturation of mass spectrometry [@HAN2008483] and soft-ionisation methods [@mas21900], started to become a cost effective alternative in identification of protease substrates at a proteome level. Therefore, the ability to rapidly develop *in silico* pipelines to support preliminary discovery of protease substrates is increasingly needed to complement the emerging experimental efforts. Here, the introduction of `substrateminer` aims to streamline the modular development of *in silico* pipelines for the preliminary discovery of substrate degradome for a target protease based on cleavage sequence consensus.
+Nevertheless, one constraining factor that confronts the rapid development of protease inhibitory therapy is the undiscriminating nature of inhibition. This is a major challenge because most proteases are involved in a complex network of biological processes and a complete inhibition of one protease can lead to unintended consequences on many other biological processes. Therefore, identifying and understanding the substrate degradome of a target protease is an important preliminary step towards inhibitory therapy. Traditionally, the identification of protease substrates is often challenging due to the prohibitive cost of large scale deployment of target-designed immuno-based biochemistry assays and scarcity of adequate animal models. Modern advancements in proteomics, particularly the maturation of mass spectrometry [@HAN2008483] and soft-ionisation methods [@Challen202100394], started to become a cost effective alternative in identification of protease substrates at a proteome level. Therefore, the ability to rapidly develop *in silico* pipelines to support preliminary discovery of protease substrates is increasingly needed to complement the emerging experimental efforts. Here, the introduction of `substrateminer` aims to streamline the modular development of *in silico* pipelines for the preliminary discovery of substrate degradome for a target protease based on cleavage sequence consensus.
 
 # Methods
 `Substrateminer` contains three main classes of functions (with a schematic depicted in \autoref{fig:workflow}), namely consensus investigation (via submodules msa and consensus), substrate mining (via submodule miner) and substrate biopathological pathway searching (via submodule pathfinder).
