+++ paper/paper.bib
@@ -247,7 +247,7 @@ pages = {185-202},
 year = {2004},
 note = {Special Issue in Honour of Alan G. Marshall},
 issn = {1387-3806},
-doi = {https://doi.org/10.1016/j.ijms.2004.02.015},
+doi = {10.1016/j.ijms.2004.02.015},
 url = {https://www.sciencedirect.com/science/article/pii/S1387380604001150},
 author = {Paul M Bigwarfe and Troy D Wood},
 keywords = {Bottom-up, Ionization, Mass fingerprint},
@@ -263,7 +263,7 @@ pages = {483-490},
 year = {2008},
 note = {Analytical Techniques/Mechanisms},
 issn = {1367-5931},
-doi = {https://doi.org/10.1016/j.cbpa.2008.07.024},
+doi = {10.1016/j.cbpa.2008.07.024},
 url = {https://www.sciencedirect.com/science/article/pii/S1367593108001178},
 author = {Xuemei Han and Aaron Aslanian and John R Yates},
 abstract = {Mass spectrometry has been widely used to analyze biological samples and has evolved into an indispensable tool for proteomics research. Our desire to understand the proteome has led to new technologies that push the boundary of mass spectrometry capabilities, which in return has allowed mass spectrometry to address an ever-increasing array of biological questions. The recent development of a novel mass spectrometer (Orbitrap) and new dissociation methods such as electron-transfer dissociation has made possible the exciting new areas of proteomic application. Although bottom-up proteomics (analysis of proteolytic peptide mixtures) remains the workhorse for proteomic analysis, middle-down and top-down strategies (analysis of longer peptides and intact proteins, respectively) should allow more complete characterization of protein isoforms and post-translational modifications. Finally, stable isotope labeling strategies have transformed mass spectrometry from merely descriptive to a tool for measuring dynamic changes in protein expression, interaction, and modification.}
@@ -275,9 +275,9 @@ title = {Protein analysis by desorption electrospray ionization mass spectrometr
 journal = {Mass Spectrometry Reviews},
 volume = {n/a},
 number = {n/a},
-pages = {},
+pages = {n/a},
 keywords = {AIMS, ambient ionization, DESI, desorption electrospray ionization, protein analysis},
-doi = {https://doi.org/10.1002/mas.21900},
+doi = {10.1002/mas.21900},
 url = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/abs/10.1002/mas.21900},
 eprint = {https://analyticalsciencejournals.onlinelibrary.wiley.com/doi/pdf/10.1002/mas.21900},
 abstract = {Abstract This review presents progress made in the ambient analysis of proteins, in particular by desorption electrospray ionization-mass spectrometry (DESI-MS). Related ambient ionization techniques are discussed in comparison to DESI-MS only to illustrate the larger context of protein analysis by ambient ionization mass spectrometry. The review describes early and current approaches for the analysis of undigested proteins, native proteins, tryptic digests, and indirect protein determination through reporter molecules. Applications to mass spectrometry imaging for protein spatial distributions, the identification of posttranslational modifications, determination of binding stoichiometries, and enzymatic transformations are discussed. The analytical capabilities of other ambient ionization techniques such as LESA and nano-DESI currently exceed those of DESI-MS for in situ surface sampling of intact proteins from tissues. This review shows, however, that despite its many limitations, DESI-MS is making valuable contributions to protein analysis. The challenges in sensitivity, spatial resolution, and mass range are surmountable obstacles and further development and improvements to DESI-MS is justified.}
@@ -323,7 +323,7 @@ volume = {48},
 number = {1},
 pages = {3.13.1-3.13.16},
 keywords = {Clustal Omega, multiple sequence alignment, iteration, HMM, hidden Markov model},
[…]
