+++ paper/paper.md
@@ -69,7 +69,7 @@ To further investigate the pathobiological relevance of novel substrates, `subst
 As of the current release, `pathfinder` submodule queries all 580 KEGG molecular pathway maps and covers 1205 eukaryotes, 9375 bacteria and 449 archaea. [@kegg_statistics_2025] Nevertheless, the diseases covered are limited to human context.
 
 # Limitations
-The current release of `substrateminer` is limited to the investigation of protease substrates based on substrate primary sequence consensus in model organisms. In cases where proteases have very loose consensus like metalloproteinase ADAM10[@CaescuBJ20090549] that is arguably best-known for its protective effects in prion diseases like mad cow in cattle, additional structural factors and contexts may need to be taken into consideration. Additionally, the efficacy of `substrateminer` investigative power expanding beyond well-studied organisms is limited by the availability of a comprehensive proteome database for underlying species. In the future, we aim to expand the scope of `substrateminer` to scrutinise substrate structural characteristics and provide a more comprehensive suite of tools for substrate investigation.
+The current release of `substrateminer` is limited to the investigation of protease substrates based on substrate primary sequence consensus in model organisms. In cases where proteases have very loose consensus like metalloproteinase ADAM10 [@CaescuBJ20090549] that is arguably best-known for its protective effects in prion diseases like mad cow in cattle, additional structural factors and contexts may need to be taken into consideration. Additionally, the efficacy of `substrateminer` investigative power expanding beyond well-studied organisms is limited by the availability of a comprehensive proteome database for underlying species. In the future, we aim to expand the scope of `substrateminer` to scrutinise substrate structural characteristics and provide a more comprehensive suite of tools for substrate investigation.
 
 # Conflict of Interests
 The author declares no conflict of interests.
