### 2025-08-31 2fcb534
Minor edits to paper
