+++ paper/paper.md
@@ -20,14 +20,14 @@ date: 29 Aug 2024
 bibliography: paper.bib
 ---
 
-# Abstract
-`Substrateminer` is a Python library that provides a suite of modular investigative methods that are based on protein sequence and cellular properties to support rapid development of *in silico* pipelines to discover and interrogate potential substrates for a target protease of interest. Three categories of methods included in `substrateminer` are consensus analysis, substrate mining, and pathological investigation, which together provide a well-suited toolkit to explore substrate repertoires also known as substrate degradome for a target protease.
+# Summary
+`Substrateminer` is a Python library that provides a suite of modular investigative methods based on protein sequence and cellular properties to support rapid development of *in silico* pipelines to discover and interrogate potential substrates for a target protease of interest. Three categories of methods are included in `substrateminer`: consensus analysis, substrate mining, and pathological investigation. Together, they provide a well-suited toolkit to explore substrate repertoires, also known as the substrate degradome, for a target protease.
 
 # Statement of need
-The ability to identify and discover novel protease *in situ* substrates is often a powerful first step towards understanding the biological functions of proteases and their implications in human health and disease. However, in traditional enzymology, substrates of a target enzyme are studied in a one-to-one relationship, arguably due to prohibitive cost involved in designing and deploying experimental assays with large quantity. Modern proteomics methods particularly the advancement in mass spectrometry, drastically reduce the cost associated with detecting substrate hydrolysis events at molecular level and now the study of substrates of a target enzyme at proteome level is no longer a distant fantasy. With these new possibilities, the lack of a modular system to enable rapid development of *in silico* pipeline to support preliminary discovery has been a bottleneck for new studies. Therefore, the introduction of `substrateminer` here aims to provide a versatile toolkit that enables fast modular development of *in silico* pipelines to facilitate the preliminary discovery that complement and accelerate the emerging experimental efforts.
+The ability to identify and discover novel protease *in situ* substrates is often a powerful first step towards understanding the biological functions of proteases and their implications in human health and disease. However, in traditional enzymology, substrates of a target enzyme are studied one at a time, arguably due to prohibitive cost involved in designing and deploying large-scale experimental assays. Modern proteomics methods, particularly the advancement in mass spectrometry, drastically reduce the cost associated with detecting substrate hydrolysis events at the molecular level, and now the study of substrates of a target enzyme at the proteome level is no longer a distant fantasy. With these new possibilities, the lack of a modular system to enable rapid development of *in silico* pipelines to support preliminary discovery has been a bottleneck for new studies. Therefore, `substrateminer` aims to provide a versatile toolkit that enables fast modular development of *in silico* pipelines to facilitate the preliminary discovery that complement and accelerate the emerging experimental efforts.
 
 # Background
-Proteolytic cleavage is a fundamental biochemical process in biology, which underpins many important biological processes that surround us from common food processing and digestion [@campbell2005biology]. Almost all proteolytic processes in nature are catalysed by enzymes and proteases are such an enzyme class that facilitates the biochemical reaction of cleaving peptide bonds. The inhibition of proteases has been a major intervention strategy in modern clinics to modulate molecular proteolytic processes to treat many physiological conditions including viral infections like HIV, hepatitis-C; metabolic dysfunctions like type-2 diabetes [@Scott2017]; and cancers [@Manasanch2017]. In the wake of the COVID-19 pandemic, proteases have emerged as a major therapeutic target for curbing fatality due to viral infections.[@ijms25158105;@v16030366;@ijms22115762]
[…]
