### 2025-09-01 babf581
fix #86 - refine language and detail in paper.md
Thx @csoneson for the comments, closes DPP4ResearchGroup#86
