### 2025-09-04 #8266 @csoneson
Thanks @manifestoso - looks good, there's only a mismatch between the licenses indicated in the Zenodo archive (MIT and CC-BY) and those in the GitHub repo (MIT). I would suggest removing the CC-BY indication from Zenodo for consistency.
